LayoutTests/http/tests/wml/go-task-get-method-accept-charset.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../../../fast/js/resources/js-test-style.css">
<script>var relativePathToLayoutTests = "../../..";</script>
<script src="../../../wml/resources/WMLTestCase.js"></script>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<h1>WML http navigation layout tests - using XHTML scripting</h1>
<p id="description"></p>
<div id="console"></div>
<script src="script-tests/go-task-get-method-accept-charset.js" charset="utf8"></script>
</body>
</html>
